Fixed bugs with converter logic

+++ bntu.vsrpp.vmilyuk.Core/Models/Currency.cs
@@ -10,7 +10,7 @@ namespace bntu.vsrpp.vmilyuk.Core.Models
         public int Cur_Code { get; set; }
         public DateTime Date { get; set; }
         public string Cur_Abbreviation { get; set; }
-        public int Cur_Scale { get; set; }
+        public float Cur_Scale { get; set; }
         public string Cur_Name { get; set; }
         public float Cur_OfficialRate { get; set; }
     }
